Move tests from integration.rs into contract.rs, add test for send_all_back, and merge query unit test into existing test
