+++ framework/contracts/native/ibc-host/src/endpoints/query.rs
@@ -23,6 +23,7 @@ pub fn query(deps: Deps, _env: Env, query: QueryMsg) -> HostResult<Binary> {
     }
     .map_err(Into::into)
 }
+
 fn config(deps: Deps) -> HostResult<ConfigResponse> {
     let state = CONFIG.load(deps.storage)?;
     Ok(ConfigResponse {
@@ -53,59 +54,4 @@ fn registered_chains(
 fn associated_client(deps: Deps, chain: String) -> HostResult<ClientProxyResponse> {
     let proxy = CHAIN_PROXIES.load(deps.storage, &ChainName::from_str(&chain)?)?;
     Ok(ClientProxyResponse { proxy })
-}
-
-#[cfg(test)]
-mod test {
-
-    #[test]
-    fn test_registered_client() {
-        use abstract_core::ibc_host::ClientProxyResponse;
-        use abstract_core::ibc_host::InstantiateMsg;
-        use abstract_core::ibc_host::QueryMsg;
-
-        use crate::contract::{execute, instantiate, query};
-        use cosmwasm_std::from_binary;
-        use cosmwasm_std::testing::mock_dependencies;
-        use cosmwasm_std::testing::mock_env;
-        use cosmwasm_std::testing::mock_info;
-        // Instantiate
-        let mut deps = mock_dependencies();
-        let info = mock_info("admin", &[]);
-        instantiate(
-            deps.as_mut(),
-            mock_env(),
-            info.clone(),
-            InstantiateMsg {
-                account_factory_address: "dummy".to_string(),
-                version_control_address: "foo".to_string(),
-                ans_host_address: "bar".to_string(),
-            },
-        )
-        .unwrap();
-
-        // Register
-        execute(
-            deps.as_mut(),
-            mock_env(),
-            info,
-            abstract_core::ibc_host::ExecuteMsg::RegisterChainProxy {
-                chain: "juno".into(),
-                proxy: "juno-proxy".to_string(),
-            },
-        )
-        .unwrap();
-
-        // Query
-        let client_name = query(
-            deps.as_ref(),
-            mock_env(),
-            QueryMsg::ClientProxy {
-                chain: "juno".to_string(),
-            },
-        )
-        .unwrap();
-        let queried_client_name: ClientProxyResponse = from_binary(&client_name).unwrap();
-        assert_eq!(queried_client_name.proxy, "juno-proxy");
-    }
 }
